feat: Refactor job notification function for better performance
